Toying with structure
Finding a lot of diferent ways to set up an app like this implemented mostly with old versions of Express, trying to set something up that makes sense and isn't overcomplicating things.  I've got the angular app placeholder running, no api set up yet but it is frameworked a little.